BeetleDashboard/requirements.txt
# Beetle Dashboard - Requirements
# Install with: pip install -r requirements.txt

Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
Flask-Compress==1.14
flask-cors==4.0.0

# Serial communication
pyserial==3.5

# HTTP client for serial reader
requests==2.31.0
